+++ GDX.Editor/UpdateProvider.cs
@@ -410,7 +410,7 @@ namespace GDX.Editor
                         Path.Combine(tempExtractFolder, "GDX-" + UpdatePackageDefinition.version),
                         targetPath);
 
-                    AssetDatabase.ImportAsset(LocalPackage.PackageAssetPath);
+                    AssetDatabase.ImportAsset(LocalPackage.PackageManagerInfo.assetPath);
                 }
                 finally
                 {
@@ -453,7 +453,7 @@ namespace GDX.Editor
                         process.WaitForExit();
 
                         // Lets force the import anyways now
-                        AssetDatabase.ImportAsset(LocalPackage.PackageAssetPath);
+                        AssetDatabase.ImportAsset(LocalPackage.PackageManagerInfo.assetPath);
                     }
                 }
                 finally